requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-filter==24.3
django-cors-headers==4.6.0
drf-spectacular==0.28.0
mysqlclient==2.2.4
celery==5.4.0
django-celery-beat==2.7.0
redis==5.2.1
Pillow==11.1.0
gunicorn==23.0.0
requests==2.32.3
